+++ src/getStepStartStates.ts
@@ -13,6 +13,7 @@ import { STEP_API_SERVICES } from './steps/service-usage';
 import { deserializeIntegrationConfig } from './utils/integrationConfig';
 import { STEP_IAM_ROLES, STEP_IAM_SERVICE_ACCOUNTS } from './steps/iam';
 import { STEP_RESOURCE_MANAGER_IAM_POLICY } from './steps/resource-manager';
+import { STEP_COMPUTE_INSTANCES, STEP_COMPUTE_DISKS } from './steps/compute';
 
 async function getEnabledServiceNames(
   config: IntegrationConfig,
@@ -89,5 +90,7 @@ export default async function getStepStartStates(
     [STEP_RESOURCE_MANAGER_IAM_POLICY]: createStepStartState(
       ServiceUsageName.RESOURCE_MANAGER,
     ),
+    [STEP_COMPUTE_DISKS]: createStepStartState(ServiceUsageName.COMPUTE),
+    [STEP_COMPUTE_INSTANCES]: createStepStartState(ServiceUsageName.COMPUTE),
   };
 }