+++ src/steps/compute/constants.ts
@@ -1,0 +1,10 @@
+export const STEP_COMPUTE_INSTANCES = 'fetch-compute-instances';
+export const STEP_COMPUTE_DISKS = 'fetch-compute-disks';
+
+export const ENTITY_CLASS_COMPUTE_DISK = 'DataStore';
+export const ENTITY_TYPE_COMPUTE_DISK = 'google_compute_disk';
+export const ENTITY_CLASS_COMPUTE_INSTANCE = 'Host';
+export const ENTITY_TYPE_COMPUTE_INSTANCE = 'google_compute_instance';
+
+export const RELATIONSHIP_TYPE_GOOGLE_COMPUTE_INSTANCE_USES_DISK =
+  'google_compute_instance_uses_disk';